feat: JobCounter에서 uint32를 usize로 변경하여 유연성 증가
- JobCounter 및 JobHandle에서 uint32를 usize로 변경하여 플랫폼 독립적인 데이터 처리 지원
- IsValid 메서드 추가로 JobHandle 유효성 확인 기능 구현
- operator bool()의 내부 로직을 IsValid 메서드로 위임
- await_suspend 및 await_ready 메서드에 주석 추가 및 코드 정리 (NOLINT 주석 포함)

+++ EngineCore/Source/Core/Concurrency/JobCounter.cpp
@@ -10,7 +10,7 @@
 
 namespace se
 {
-JobCounter::JobCounter(uint32 in_initial_count)
+JobCounter::JobCounter(usize in_initial_count)
     : count(in_initial_count)
 {
     if (in_initial_count == 0)
@@ -39,7 +39,7 @@ JobCounter::~JobCounter()
 
 void JobCounter::Decrement()
 {
-    const uint32 prev = count.fetch_sub(1, std::memory_order_acq_rel);
+    const usize prev = count.fetch_sub(1, std::memory_order_acq_rel);
     SE_ASSERT(prev > 0, "Decrement called on already-zero counter");
 
     if (prev == 1)
@@ -55,7 +55,7 @@ bool JobCounter::IsComplete() const
     return count.load(std::memory_order_acquire) == 0;
 }
 
-uint32 JobCounter::GetCount() const
+usize JobCounter::GetCount() const
 {
     return count.load(std::memory_order_acquire);
 }
